refactor: 폴더 구조 수정 및 컴포넌트 분류
